+++ CHANGELOG.md
@@ -14,6 +14,8 @@
 
 - CHANGED: Updated whois.gg parsers to the new response format.
 
+- CHANGED: Updated whois.pnina.ps parsers to the new response format.
+
 
 ## Release 3.2.0